Refactor dsl
