+++ src/main/scala/outwatch/dom/task.scala
@@ -1,0 +1,11 @@
+package outwatch.dom
+
+import cats.effect.Effect
+import monix.eval.Task
+import monix.execution.Scheduler
+
+object task {
+  implicit def dsl(implicit ec: Scheduler): OutwatchDsl[Task] = new OutwatchDsl[Task] {
+    implicit val effectF: Effect[Task] = Task.catsEffect(ec)
+  }
+}
